requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
deep-translator==1.11.4
fonttools==4.61.1
future==1.0.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.0
packaging==25.0
pillow==12.0.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
soupsieve==2.8.1
sv_ttk==2.6.1
ttkbootstrap==1.19.0
typing_extensions==4.15.0
urllib3==2.6.3
